re PR target/54639 (mn10300_expand_epilogue: signed / unsigned comparison)
        PR target/54639
        * config/mn10300/mn10300.c (mn10300_expand_epilogue): Avoid offset
        type promotion to unsigned.

From-SVN: r196289

+++ gcc/config/mn10300/mn10300.c
@@ -1077,7 +1077,7 @@ mn10300_expand_epilogue (void)
 	      /* Insn: add size + 4 * num_regs_to_save
 				+ reg_save_bytes - 252,sp.  */
 	      this_strategy_size = SIZE_ADD_SP (size + 4 * num_regs_to_save
-						+ reg_save_bytes - 252);
+						+ (int) reg_save_bytes - 252);
 	      /* Insn: fmov (##,sp),fs#, fo each fs# to be restored.  */
 	      this_strategy_size += SIZE_FMOV_SP (252 - reg_save_bytes
 						  - 4 * num_regs_to_save,
